+++ src/Tests/Analyzers.Tests/RCS1248NormalizeNullCheckTests.cs
@@ -211,4 +211,27 @@ class C
 }
 ", options: Options.AddConfigOption(ConfigOptionKeys.NullCheckStyle, ConfigOptionValues.NullCheckStyle_PatternMatching));
     }
+
+    [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.NormalizeNullCheck)]
+    public async Task TestNoDiagnostic_ExpressionTree2()
+    {
+        await VerifyNoDiagnosticAsync("""
+using System.Linq;
+
+class C
+{
+    void M()
+    {
+        var _ = from x in new[] { "a" }.AsQueryable()
+                join yy in new[] { "b" }.AsQueryable()
+                on x equals yy into y
+                from yy in y.DefaultIfEmpty()
+                select new
+                {
+                    A = yy != null
+                };
+    }
+}
+""", options: Options.AddConfigOption(ConfigOptionKeys.NullCheckStyle, ConfigOptionValues.NullCheckStyle_PatternMatching));
+    }
 }